Собирается под mac

+++ tests/single_thread/TestWsPayloadLength.cpp
@@ -9,7 +9,7 @@ BOOST_AUTO_TEST_SUITE(SuiteWsPayloadLength)
 BOOST_AUTO_TEST_CASE(TestRead) {
 	Buffer buffer {0x7e, 0x08, 0x40, 0x8b, 0x00, 0x0b};
 	auto it = buffer.begin();
-	size_t integer;
+	uint64_t integer;
 	BOOST_REQUIRE_NO_THROW(it = ReadWsPayloadLength(it, buffer.end(), &integer));
 	BOOST_REQUIRE(integer == 2112);
 	BOOST_REQUIRE_NO_THROW(it = ReadWsPayloadLength(it, buffer.end(), &integer));
